+++ core/src/edu/cornell/gdiac/sync/ISynced.java
@@ -6,7 +6,7 @@ public interface ISynced {
      * Assume in common time: 4/4
      * Ex: 1 represents a pulse every quarter note, .5 represents a pulse every half note
      */
-    public int getBeat();
+    public float getBeat();
 
     /**
      * Called when an update should occur on the beat
